Revert "Reduce comp. time of portable_storage_template_helper.h"
This reverts commit 673c6d2d11a4ae639c43bbf40e31e808932fd625.
