+++ contrib/epee/include/storages/portable_storage_template_helper.h
@@ -29,7 +29,7 @@
 #include <string>
 
 #include "byte_slice.h"
-#include "parserse_base_utils.h" /// TODO: (mj-xmr) This will be reduced in an another PR
+#include "parserse_base_utils.h"
 #include "portable_storage.h"
 #include "file_io_utils.h"
 #include "span.h"
