+++ src/common/i18n.cpp
@@ -35,8 +35,6 @@
 #include "common/i18n.h"
 #include "translation_files.h"
 
-#include <boost/system/error_code.hpp>
-#include <boost/filesystem.hpp>
 #include <algorithm>
 
 #undef MONERO_DEFAULT_LOG_CATEGORY
