+++ src/cryptonote_basic/miner.cpp
@@ -43,7 +43,6 @@
 #include "string_tools.h"
 #include "storages/portable_storage_template_helper.h"
 #include "boost/logic/tribool.hpp"
-#include <boost/filesystem.hpp>
 
 #ifdef __APPLE__
   #include <sys/times.h>
