+++ src/daemon/command_parser_executor.cpp
@@ -30,7 +30,6 @@
 #include "common/command_line.h"
 #include "net/parse.h"
 #include "daemon/command_parser_executor.h"
-#include <boost/filesystem.hpp>
 #include <boost/algorithm/string/predicate.hpp>
 
 #undef MONERO_DEFAULT_LOG_CATEGORY
