+++ src/wallet/api/pending_transaction.cpp
@@ -40,7 +40,6 @@
 #include <vector>
 #include <sstream>
 #include <boost/format.hpp>
-#include <boost/filesystem.hpp>
 
 using namespace std;
