+++ src/wallet/api/wallet.cpp
@@ -45,8 +45,10 @@
 #include <sstream>
 #include <unordered_map>
 
+#ifdef WIN32
 #include <boost/locale.hpp>
 #include <boost/filesystem.hpp>
+#endif
 
 using namespace std;
 using namespace cryptonote;
